Split schemas to files
